refactor(capabilities): relocate mat4_apply to aether-kinds and test_echo under rpc::server
Two non-capability residents leave the crate root:

- mat4_apply (the sole first-party #[transform]) moves to aether-kinds,
  co-located with the Mat4Apply kind it consumes. Both link consumers
  (aether-substrate-bundle, aether-mcp) already depend on aether-kinds,
  so the link-time inventory submission still populates the substrate
  TransformRegistry and describe_transforms. aether-capabilities no
  longer needs its aether-math dependency (mat4_apply was its only user).

- test_echo (the far-end receiver of an RPC Call) moves from the crate
  root to crate::rpc::server::test_echo as a #[cfg(test)] pub mod. Both
  consumers are RPC-server round-trips: rpc::server's own tests, and the
  engine::proxy test that forwards onto the same Call path through a
  booted RpcServerCapability. The cfg(test) pub never reaches the cap's
  shipped surface while staying reachable from the sibling engine::proxy.
