+++ crates/aether-capabilities/Cargo.toml
@@ -108,10 +108,6 @@ aether-actor = { path = "../aether-actor" }
 aether-codec = { path = "../aether-codec" }
 aether-data = { path = "../aether-data" }
 aether-kinds = { path = "../aether-kinds" }
-# Issue 1464: the `mat4_apply` first-party transform rebuilds a `Mat4`
-# / `Vec4` from the wire `[f32; N]` arrays. Always-on + target-agnostic
-# (no_std), so it rides the wasm-header-only build too.
-aether-math = { path = "../aether-math" }
 # `alloc` rides along for the `rpc::wire` + `trace_walk` modules (folded
 # in per ADR-0124), whose owned-`String`/`Vec` wire types need serde's
 # alloc impls even in the marker-only isolated build.
